requirements.txt
Appium-Python-Client==4.1.0
imageio==2.35.1
numpy==2.1.1
PyAutoGUI==0.9.54
pytest==8.3.3
pytest-assume==2.4.3
selenium==4.24.0
